refactor(packages): fold capture packages into core and viewer
Avoid npm package sprawl before 1.0.0: `@pascal-app/capture-protocol`
becomes the `@pascal-app/core/capture` subpath and
`@pascal-app/capture-viewer` becomes `@pascal-app/viewer/capture` (plus
`@pascal-app/viewer/capture/preview`), so the release ships seven
packages: core, viewer, editor, nodes, mcp, ifc-converter, cli.

Neither package was ever published to npm, so no npm consumer migrates.
The protocol code is pure zod/TS, so core keeps its no-Three.js layer
rule; the runtime and its reference layers keep viewer's existing peers
and now reach viewer internals through relative imports instead of a
self-referential `@pascal-app/viewer` specifier.
