--- packages/capture-protocol/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@pascal-app/capture-protocol",
-  "version": "1.0.0-beta.4",
-  "description": "Transport-neutral capture-session manifests and live stream sources for Pascal",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "react-native": "./src/index.ts",
-      "browser": "./src/index.ts",
-      "import": "./dist/index.js",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "src",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsc --build",
-    "dev": "tsgo --build --watch",
-    "test": "bun test src",
-    "prepublishOnly": "npm run build"
-  },
-  "dependencies": {
-    "zod": ">=4.5.4 <4.6"
-  },
-  "devDependencies": {
-    "@pascal/typescript-config": "*",
-    "@types/bun": "^1.3.0",
-    "typescript": "6.0.3"
-  },
-  "keywords": [
-    "3d",
-    "capture",
-    "point-cloud",
-    "sensor-fusion",
-    "streaming"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/pascalorg/editor.git",
-    "directory": "packages/capture-protocol"
-  },
-  "license": "MIT",
-  "homepage": "https://github.com/pascalorg/editor/tree/main/packages/capture-protocol#readme",
-  "bugs": "https://github.com/pascalorg/editor/issues"
-}
