--- packages/capture-protocol/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@pascal/typescript-config/react-library.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "noEmit": false,
-    "composite": true,
-    "incremental": true,
-    "types": ["bun"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "**/*.test.ts"]
-}
