--- packages/capture-viewer/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@pascal/typescript-config/react-library.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "noEmit": false,
-    "composite": true,
-    "incremental": true,
-    "types": ["bun"]
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.test.tsx"],
-  "references": [{ "path": "../capture-protocol" }, { "path": "../core" }, { "path": "../viewer" }]
-}
