+++ packages/core/README.md
@@ -23,6 +23,9 @@ npm install react three @react-three/fiber @react-three/drei
 - **Spatial Grid** - Collision detection and placement validation
 - **Event Bus** - Typed event emitter for inter-component communication
 - **Asset Storage** - IndexedDB-based file storage for user-uploaded assets
+- **Capture Contracts** (`@pascal-app/core/capture`) - Versioned capture-session manifests,
+  normalized stream descriptors, packet headers, and transport-neutral static/live `CaptureSource`
+  implementations
 
 ## Usage
 
@@ -77,6 +80,26 @@ await loadPlugin(builtinPlugin)
 Load the plugin before mounting `@pascal-app/viewer`. See the
 [`@pascal-app/viewer` quick start](https://github.com/pascalorg/editor/tree/main/packages/viewer#usage)
 for a React example.
+
+## Capture Sessions
+
+Capture contracts are a self-contained subpath — no React, no Three.js, no prescribed transport:
+
+```typescript
+import { createHttpCaptureSource, type CaptureSessionLocator } from '@pascal-app/core/capture'
+
+const locator: CaptureSessionLocator = {
+  sessionId: 'capture_123',
+  manifestUrl: '/api/captures/capture_123/manifest',
+}
+
+const source = createHttpCaptureSource(locator, { credentials: 'include' })
+const descriptor = await source.describe()
+```
+
+For live producers, use `PushCaptureSource` directly or implement `CaptureSource.subscribe()` with
+the same descriptor and packet event contract. The reference renderers that consume these sources
+ship in [`@pascal-app/viewer/capture`](https://github.com/pascalorg/editor/tree/main/packages/viewer#capture-sessions).
 
 ## License
