+++ packages/core/src/capture/index.ts
