+++ packages/core/src/capture/schema.test.ts
