+++ packages/core/src/capture/schema.ts
