+++ packages/core/src/capture/source.test.ts
