+++ packages/core/src/capture/source.ts
