+++ packages/viewer/src/capture/layer-visibility.test.ts
