+++ packages/viewer/src/capture/layers/clay-matcap.ts
