+++ packages/viewer/src/capture/layers/device-motion-layer.tsx
