+++ packages/viewer/src/capture/layers/surface-mesh-layer.tsx
