+++ packages/viewer/src/capture/preview.ts
