+++ packages/viewer/src/capture/stream-rendering.test.ts
