+++ packages/viewer/src/capture/surface-mesh-layer.test.ts
