+++ packages/viewer/src/capture/trajectory.test.ts
